+++ src/pages/api/post/[id].ts
@@ -24,6 +24,18 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse) 
     }
   } else if (req.method === 'DELETE') {
     try {
+      const session = await getSession({ req })
+      if (!session) return res.status(401).end();
+      const postId = req.query.id
+      const postQ = await prisma.post.findUnique({
+        select: {
+          author: {
+            select: { id: true },
+          },
+        },
+        where: { id: Number(postId) }
+      })
+      if (session?.user.role != 'ADMIN') if (session?.user.id != postQ?.author.id) return res.status(401).end();
       const post = await prisma.post.delete({
         where: { id: Number(postId) },
       })
